+++ src/constants/Validation/patron-notice-template.js
@@ -1,0 +1,6 @@
+/* eslint-disable import/prefer-default-export */
+export const PATRON_NOTICE_PATH = {
+  CATEGORY: 'category',
+  LOCALIZED_TEMPLATES_EN_HEADER: 'localizedTemplates.en.header',
+  LOCALIZED_TEMPLATES_EN_BODY: 'localizedTemplates.en.body',
+};