+++ CHANGELOG.md
@@ -38,6 +38,10 @@
 * Add RTL/Jest testing for `LoanHistoryModel` in `src/settings/Models`. Refs UICIRC-802.
 * Add RTL/Jest testing for `RulesField` component in `src/settings/lib/RuleEditor`. Refs UICIRC-826.
 * Add RTL/Jest testing for `src/settings/Validation/checkout-settings` folder. Refs UICIRC-805.
+* Cover `RequestPolicy` by RTL/jest tests. Refs UICIRC-801.
+* Add RTL/Jest testing for `RulesForm` component in `src/settings/lib/RuleEditor`. Refs UICIRC-825.
+* Add RTL/Jest testing for `src/settings/Models/LostItemFeePolicy` folder. Refs UICIRC-799.
+* Add RTL/Jest testing for `patron-notice-template` and `request-policy` components in `src/settings/Validation`. Refs UICIRC-812.
 * UI tests replacement with RTL/Jest for `NoticePolicy` folder in `Models`. UICIRC-800.
 
 ## [7.0.3](https://github.com/folio-org/ui-circulation/tree/v7.0.3) (2022-04-11)